HSEARCH-3773 Move some runtime bridge-related classes to a more fitting package
