+++ mapper/javabean/src/main/java/org/hibernate/search/mapper/javabean/session/impl/JavaBeanSessionIndexedTypeContext.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.javabean.session.impl;
 
-import org.hibernate.search.mapper.pojo.bridge.mapping.spi.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.runtime.spi.IdentifierMapping;
 import org.hibernate.search.mapper.pojo.model.spi.PojoRawTypeIdentifier;
 
 /**
