+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/runtime/impl/IdentifierMappingImplementor.java
@@ -4,11 +4,11 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.mapping.impl;
+package org.hibernate.search.mapper.pojo.bridge.runtime.impl;
 
 import java.util.function.Supplier;
 
-import org.hibernate.search.mapper.pojo.bridge.mapping.spi.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.runtime.spi.IdentifierMapping;
 import org.hibernate.search.mapper.pojo.mapping.context.spi.AbstractPojoBackendMappingContext;
 
 /**
