+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/runtime/impl/PojoValueBridgeFromDocumentFieldValueConverter.java
@@ -4,18 +4,18 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.binding.impl;
+package org.hibernate.search.mapper.pojo.bridge.runtime.impl;
 
 import org.hibernate.search.engine.backend.types.converter.FromDocumentFieldValueConverter;
 import org.hibernate.search.engine.backend.types.converter.runtime.FromDocumentFieldValueConvertContext;
 import org.hibernate.search.mapper.pojo.bridge.ValueBridge;
 
-final class PojoValueBridgeFromDocumentFieldValueConverter<F, V>
+public final class PojoValueBridgeFromDocumentFieldValueConverter<F, V>
 		implements FromDocumentFieldValueConverter<F, V> {
 
 	private final ValueBridge<V, F> bridge;
 
-	PojoValueBridgeFromDocumentFieldValueConverter(ValueBridge<V, F> bridge) {
+	public PojoValueBridgeFromDocumentFieldValueConverter(ValueBridge<V, F> bridge) {
 		this.bridge = bridge;
 	}
