+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/runtime/impl/PojoValueBridgeToDocumentFieldValueConverter.java
@@ -4,19 +4,19 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.binding.impl;
+package org.hibernate.search.mapper.pojo.bridge.runtime.impl;
 
 
 import org.hibernate.search.engine.backend.types.converter.ToDocumentFieldValueConverter;
 import org.hibernate.search.engine.backend.types.converter.runtime.ToDocumentFieldValueConvertContext;
 import org.hibernate.search.mapper.pojo.bridge.ValueBridge;
 
-final class PojoValueBridgeToDocumentFieldValueConverter<V, F>
+public final class PojoValueBridgeToDocumentFieldValueConverter<V, F>
 		implements ToDocumentFieldValueConverter<V, F> {
 
 	private final ValueBridge<V, F> bridge;
 
-	PojoValueBridgeToDocumentFieldValueConverter(ValueBridge<V, F> bridge) {
+	public PojoValueBridgeToDocumentFieldValueConverter(ValueBridge<V, F> bridge) {
 		this.bridge = bridge;
 	}
