+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/bridge/runtime/impl/RoutingKeyProvider.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.bridge.mapping.impl;
+package org.hibernate.search.mapper.pojo.bridge.runtime.impl;
 
 import java.util.function.Supplier;
