+++ mapper/pojo-base/src/main/java/org/hibernate/search/mapper/pojo/mapping/building/impl/PojoIdentityMappingCollectorImpl.java
@@ -12,11 +12,11 @@ import java.util.Optional;
 import org.hibernate.search.engine.mapper.mapping.building.spi.IndexedEntityBindingContext;
 import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundIdentifierBridge;
 import org.hibernate.search.mapper.pojo.bridge.binding.impl.BoundRoutingKeyBridge;
-import org.hibernate.search.mapper.pojo.bridge.mapping.impl.IdentifierMappingImplementor;
-import org.hibernate.search.mapper.pojo.bridge.mapping.impl.PropertyIdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.mapping.impl.ProvidedStringIdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.mapping.impl.RoutingKeyBridgeRoutingKeyProvider;
-import org.hibernate.search.mapper.pojo.bridge.mapping.impl.RoutingKeyProvider;
+import org.hibernate.search.mapper.pojo.bridge.runtime.impl.IdentifierMappingImplementor;
+import org.hibernate.search.mapper.pojo.bridge.runtime.impl.PropertyIdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.runtime.impl.ProvidedStringIdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.runtime.impl.RoutingKeyBridgeRoutingKeyProvider;
+import org.hibernate.search.mapper.pojo.bridge.runtime.impl.RoutingKeyProvider;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.IdentifierBinder;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.RoutingKeyBinder;
 import org.hibernate.search.mapper.pojo.logging.impl.Log;
